requirements.txt
fastapi
uvicorn[standard]
aiofiles
google-generativeai

